Reorganize UI code into separate files, upgrade React
