+++ datasource.jsx
@@ -1,0 +1,26 @@
+/* global React,App*/
+
+App.Datasource = React.createClass({
+
+    onChangeDataLib: function(ev) {
+      window.location.hash = "#"+ ev.target.value;
+      window.location.reload();
+    },
+
+    render: function() {
+        return (
+          <header className="clearfix">
+            <span className="pull-right">
+              Data source:
+              <select
+                value={this.props.datalib} 
+                onChange={this.onChangeDataLib}>
+                {this.props.datalibs.map(function(lib){
+                    return <option key={lib}>{lib}</option>;
+                })}
+              </select>
+            </span>
+          </header>
+        );
+    }
+});
