+++ ffi.js
@@ -9,6 +9,6 @@ window.lua_ffi = {
 		return [];
 	},
 	push_to_array: function(arr, value)  {
-		arr.push(value)
+		arr.push(value);
 	}
 }
